+++ src/node-lib/curriculum-api-2023/shared.schema.ts
@@ -180,7 +180,7 @@ export const baseLessonOverviewSchema = z.object({
   isWorksheetLandscape: z.boolean().optional().nullable(),
   hasDownloadableResources: z.boolean(),
   hasCopyrightMaterial: z.boolean().optional().nullable(),
-  expired: z.boolean().optional().nullable(),
+  expired: z.boolean().nullable(),
   starterQuiz: lessonOverviewQuizData,
   exitQuiz: lessonOverviewQuizData,
   videoTitle: z.string().nullish(),
@@ -188,7 +188,7 @@ export const baseLessonOverviewSchema = z.object({
 });
 export type LessonBase = z.infer<typeof baseLessonOverviewSchema>;
 
-const lessonDownloadsListSchema = z.array(
+export const lessonDownloadsListSchema = z.array(
   z.object({
     exists: z.boolean().nullable(),
     type: z.enum([
@@ -218,6 +218,7 @@ export const baseLessonDownloadsSchema = z.object({
   lessonTitle: z.string(),
   downloads: lessonDownloadsListSchema,
   hasDownloadableResources: z.boolean(),
+  expired: z.boolean().nullable(),
 });
 
 export const lessonListSchema = z.array(
@@ -226,7 +227,7 @@ export const lessonListSchema = z.array(
     lessonTitle: z.string(),
     description: z.string(),
     pupilLessonOutcome: z.string().nullish(),
-    expired: z.boolean(),
+    expired: z.boolean().nullable(),
     quizCount: z.number().nullish(),
     videoCount: z.number().nullish(),
     presentationCount: z.number().nullish(),
@@ -236,3 +237,30 @@ export const lessonListSchema = z.array(
     lessonCohort: z.string().nullish(),
   }),
 );
+
+export const legacyAssetObjectSchema = z
+  .object({
+    google_drive: z.object({
+      id: z.string(),
+      url: z.string(),
+    }),
+    google_drive_downloadable_version: z
+      .object({
+        id: z.string(),
+        url: z.string(),
+      })
+      .nullish(),
+  })
+  .nullish();
+
+export const lessonShareResourceSchema = z.object({
+  exists: z.boolean().nullable(),
+  type: z.enum([
+    "intro-quiz-questions",
+    "exit-quiz-questions",
+    "worksheet-pdf",
+    "video",
+  ]),
+  label: z.string(),
+  metadata: z.string().nullable(),
+});