+++ src/node-lib/curriculum-api/index.ts
@@ -236,6 +236,7 @@ export const lessonShareSchema = z.intersection(
     lessonSlug: z.string(),
     lessonTitle: z.string(),
     shareableResources: z.array(lessonShareListSchema),
+    expired: z.boolean().nullable(),
   }),
 );
 
@@ -569,7 +570,6 @@ const curriculumApi = {
   lessonListing: async (...args: Parameters<typeof sdk.lessonListing>) => {
     const res = await sdk.lessonListing(...argsRemoveLegacySlugSuffix(args));
     const { units = [], lessons = [] } = transformMVCase(res);
-
     const unit = getFirstResultOrWarnOrFail()({
       results: units,
     });
@@ -597,6 +597,7 @@ const curriculumApi = {
 
     return lessonShareSchema.parse({
       ...share,
+      expired: false,
       programmeSlug: addLegacySlugSuffix(share.programmeSlug),
       isLegacy: true,
       lessonCohort: LEGACY_COHORT,
@@ -616,6 +617,7 @@ const curriculumApi = {
 
     return lessonDownloadsSchema.parse({
       ...download,
+      expired: false,
       nextLessons,
       programmeSlug: addLegacySlugSuffix(download.programmeSlug),
       isLegacy: true,
@@ -638,6 +640,7 @@ const curriculumApi = {
       },
       {
         ...downloads[0],
+        expired: false,
         pathways: [],
         isLegacy: true,
         hasDownloadableResources: true,